Formatting updates for readability and refactoring classes for ease of navigation
