+++ GauchoGrub/app/src/main/java/edu/ucsb/cs/cs190i/gauchogrub/gauchogrub/fragments/SwipesFragment.java
@@ -1,32 +1,29 @@
-package edu.ucsb.cs.cs190i.gauchogrub.gauchogrub;
+package edu.ucsb.cs.cs190i.gauchogrub.gauchogrub.fragments;
 
 import android.content.Context;
-import android.content.SharedPreferences;
 import android.net.Uri;
 import android.os.Bundle;
 import android.support.v4.app.Fragment;
-import android.support.v7.widget.LinearLayoutManager;
-import android.support.v7.widget.RecyclerView;
 import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
-
-import java.util.concurrent.ExecutorService;
-import java.util.concurrent.Executors;
+import android.webkit.WebView;
 
 import butterknife.Bind;
 import butterknife.ButterKnife;
+import edu.ucsb.cs.cs190i.gauchogrub.gauchogrub.activities.MainActivity;
+import edu.ucsb.cs.cs190i.gauchogrub.gauchogrub.R;
 
 
 /**
  * A simple {@link Fragment} subclass.
  * Activities that contain this fragment must implement the
- * {@link FavoritesFragment.OnFragmentInteractionListener} interface
+ * {@link SwipesFragment.OnFragmentInteractionListener} interface
  * to handle interaction events.
- * Use the {@link FavoritesFragment#newInstance} factory method to
+ * Use the {@link SwipesFragment#newInstance} factory method to
  * create an instance of this fragment.
  */
-public class FavoritesFragment extends Fragment {
+public class SwipesFragment extends Fragment {
     // TODO: Rename parameter arguments, choose names that match
     // the fragment initialization parameters, e.g. ARG_ITEM_NUMBER
     private static final String ARG_PARAM1 = "param1";
@@ -38,15 +35,10 @@ public class FavoritesFragment extends Fragment {
 
     private OnFragmentInteractionListener mListener;
 
-    @Bind(R.id.FavoritesFragment_recyclerView)
-    RecyclerView recyclerView;
+    @Bind(R.id.SwipesFragment_webView)
+    public WebView webView;
 
-    private String diningCommon;
-    private FavoriteRecyclerAdapter favoriteRecyclerAdapter;
-    ExecutorService executorService;
-
-
-    public FavoritesFragment() {
+    public SwipesFragment() {
         // Required empty public constructor
     }
 
@@ -56,11 +48,11 @@ public class FavoritesFragment extends Fragment {
      *
      * @param param1 Parameter 1.
      * @param param2 Parameter 2.
-     * @return A new instance of fragment FavoritesFragment.
+     * @return A new instance of fragment SwipesFragment.
      */
     // TODO: Rename and change types and number of parameters
-    public static FavoritesFragment newInstance(String param1, String param2) {
-        FavoritesFragment fragment = new FavoritesFragment();
+    public static SwipesFragment newInstance(String param1, String param2) {
+        SwipesFragment fragment = new SwipesFragment();
         Bundle args = new Bundle();
         args.putString(ARG_PARAM1, param1);
         args.putString(ARG_PARAM2, param2);
@@ -75,27 +67,23 @@ public class FavoritesFragment extends Fragment {
             mParam1 = getArguments().getString(ARG_PARAM1);
             mParam2 = getArguments().getString(ARG_PARAM2);
         }
-        SharedPreferences sharedPreferences = getActivity().getSharedPreferences(getString(R.string.MainActivity_dining_common_shared_prefs), Context.MODE_PRIVATE);
-        diningCommon = sharedPreferences.getString(MainActivity.STATE_CURRENT_DINING_COMMON,
-                sharedPreferences.getString(getString(R.string.pref_key_default_dining_common),
-                        getString(R.string.DLG)));
     }
 
     @Override
     public void onActivityCreated(Bundle savedInstanceState) {
         super.onActivityCreated(savedInstanceState);
         MainActivity activity = (MainActivity) getActivity();
-        activity.fab.show();
-        activity.updateAppBarTitle(getString(R.string.FavoritesFragment_app_bar_title), true);
+        activity.fab.hide();
+        activity.updateAppBarTitle(getString(R.string.SwipesFragment_app_bar_title), false);
     }
 
     @Override
     public View onCreateView(LayoutInflater inflater, ViewGroup container,
                              Bundle savedInstanceState) {
         // Inflate the layout for this fragment
-        View view = inflater.inflate(R.layout.fragment_favorites, container, false);
+        View view = inflater.inflate(R.layout.fragment_swipes, container, false);
         ButterKnife.bind(this, view);
-        setRecyclerAdapter(diningCommon);
+        webView.loadUrl(getString(R.string.swipes_url));
         return view;
     }
 
@@ -104,29 +92,6 @@ public class FavoritesFragment extends Fragment {
         if (mListener != null) {
             mListener.onFragmentInteraction(uri);
         }
-    }
-
-    public void switchDiningCommon(String diningCommon) {
-        this.diningCommon = diningCommon;
-        setRecyclerAdapter(diningCommon);
-    }
-
-    private void setRecyclerAdapter(String diningCommon) {
-        if(favoriteRecyclerAdapter != null) {
-            recyclerView.setRecycledViewPool(new RecyclerView.RecycledViewPool());
-            favoriteRecyclerAdapter.close();
-            executorService.shutdownNow();
-        }
-        // Create new adapter and executor
-        favoriteRecyclerAdapter = new FavoriteRecyclerAdapter(diningCommon, getContext());
-        executorService = Executors.newSingleThreadExecutor();
-        // Set executor and adapter
-        favoriteRecyclerAdapter.setExecutor(executorService);
-        recyclerView.setAdapter(favoriteRecyclerAdapter);
-        // Set layout manager
-        recyclerView.setLayoutManager(new LinearLayoutManager(getContext()));
-        // Start async query
-        favoriteRecyclerAdapter.queryAsync();
     }
 
     @Override
