+++ GauchoGrub/app/src/main/res/layout/activity_splash.xml
@@ -3,7 +3,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:background="#0099cc"
-    tools:context="edu.ucsb.cs.cs190i.gauchogrub.gauchogrub.SplashActivity">
+    tools:context=".activities.SplashActivity">
 
     <!-- The primary full-screen view. This can be replaced with whatever view
          is needed to present your content, e.g. VideoView, SurfaceView,
