+++ GauchoGrub/app/src/main/res/layout/content_main.xml
@@ -5,7 +5,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     app:layout_behavior="@string/appbar_scrolling_view_behavior"
-    tools:context="edu.ucsb.cs.cs190i.gauchogrub.gauchogrub.MainActivity"
+    tools:context=".activities.MainActivity"
     tools:showIn="@layout/app_bar_main"
     android:id="@+id/MainActivity_fragmentWrapper">
