+++ GauchoGrub/app/src/main/res/layout/fragment_favorites.xml
@@ -2,7 +2,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context="edu.ucsb.cs.cs190i.gauchogrub.gauchogrub.FavoritesFragment">
+    tools:context=".fragments.FavoritesFragment">
 
     <android.support.v7.widget.RecyclerView xmlns:android="http://schemas.android.com/apk/res/android"
         xmlns:app="http://schemas.android.com/apk/res-auto"
@@ -12,6 +12,6 @@
         android:layout_width="match_parent"
         android:layout_height="match_parent"
         app:layoutManager="LinearLayoutManager"
-        tools:context="edu.ucsb.cs.cs190i.gauchogrub.gauchogrub.FavoritesFragment"
+        tools:context=".fragments.FavoritesFragment"
         tools:listitem="@layout/fragment_favorite"/>
 </FrameLayout>
