+++ GauchoGrub/app/src/main/res/layout/fragment_schedule.xml
@@ -6,7 +6,7 @@
     android:paddingLeft="@dimen/activity_horizontal_margin"
     android:paddingRight="@dimen/activity_horizontal_margin"
     android:paddingTop="@dimen/activity_vertical_margin"
-    tools:context="edu.ucsb.cs.cs190i.gauchogrub.gauchogrub.ScheduleFragment">
+    tools:context=".fragments.ScheduleFragment">
 
     <!-- TODO: Update blank fragment layout -->
     <TextView
