+++ GauchoGrub/app/src/main/res/layout/fragment_swipes.xml
@@ -2,7 +2,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context="edu.ucsb.cs.cs190i.gauchogrub.gauchogrub.SwipesFragment">
+    tools:context=".fragments.SwipesFragment">
 
     <WebView
         android:layout_width="match_parent"
